Add files via upload
I wrote a solution to a competitive programming question, however my solutions seems rather long and slow compared to others. Tips are much appreciated.

Solution to the question Carnival_General.
The setting for the question is that we have some soldiers that are to line up next to each other. Some of them don't like each other, we are therefore given a list. Soldiers cannot stand next to someone that is on the right hand side of their list, where middle index is counted as safe. People of the left hand side of the list they like. 
For the input we are first given N, the amount of soldiers. We can assume soldier 0 likes everyone. Next we are given N-1 lines of input indicating the solders soldier k likes and dislikes. 
We are to print the list of soldiers as they are able to stand.
An example input would be:
5
0 1
0 1 2
0 1 2 3
with output:
2 0 4 1 3